LRQA-38719 Make Job into an interface, and have BaseJob implement Job
